extractor-python/requirements.txt
# extractor-python/requirements.txt
fastapi
uvicorn
pillow
pymupdf